// frontend/libs/firdaymake-up/questionnaire/src/lib/models/questionnaire.model.ts
export interface Questionnaire {
  readonly id: string;
  readonly name: string;
  readonly email: string;
  readonly phoneNumber: string;
  readonly service: string;
  readonly allergy: boolean;
  readonly allergyIngredients?: string;
  readonly skinChanges: boolean;
  readonly lenses: boolean;
  readonly pores: boolean;
  readonly medicines: boolean;
  readonly skinDiseases: boolean;
  readonly cosmeticsIngredients: boolean;
  readonly whichIngredients?: string;
  readonly diseaseOne: boolean;
  readonly diseaseTwo: boolean;
  readonly diseaseThree: boolean;
  readonly skinShiny: boolean;
  readonly skinType: string;
  readonly expectedEffect?: string;
  readonly hairdo?: string;
  readonly makeUp?: string;
  readonly cream?: string;
  readonly useAppearance: boolean;
  readonly selectedDate: Date;
  readonly selectedHour: string;
  readonly userId: string;
  readonly isClosed: boolean;
}